tutorial-validation/requirements.txt
streamlit>=1.40
qdrant-client>=1.12
fastembed>=0.4
openai>=1.50
anthropic>=0.40
sentence-transformers>=3.0
datasets>=3.0
ragas>=0.2
deepeval>=2.0
langchain-openai>=0.2
langchain-anthropic>=0.2
pandas>=2.2
numpy>=2.0
plotly>=5.24
python-dotenv>=1.0
tqdm>=4.66
pytest>=8.0
tabulate>=0.9
scipy>=1.14
